refactoring reducer
