+++ src/App.js
@@ -1,6 +1,5 @@
 import './App.css';
 import Likes from "./components/Likes";
-import store from "./redux/store";
 
 const App = () => {
   return (
@@ -9,7 +8,7 @@ const App = () => {
         <div className="card">
           <div className="card-image">
             <img src="./sea.jpg" alt="surfing"/>
-            <Likes store={store}/>
+            <Likes />
           </div>
         </div>
       </div>
