+++ src/index.js
@@ -2,8 +2,15 @@ import React from 'react';
 import ReactDOM from 'react-dom';
 import './index.css';
 import App from './App';
+import {createStore} from "redux";
+import {rootReducer} from "./redux/rootReducer";
+import {Provider} from "react-redux";
+
+const store = createStore(rootReducer)
 
 ReactDOM.render(
-  <App />,
-  document.getElementById('root')
+    <Provider store={store}>
+      <App/>
+    </Provider>,
+    document.getElementById('root')
 );
