+++ src/redux/rootReducer.js
@@ -1,0 +1,6 @@
+import {combineReducers} from "redux";
+import {likesReducer} from "./likesReducer";
+
+export const rootReducer = combineReducers({
+  likesReducer
+})
