+++ src/redux/types.js
@@ -1,0 +1,2 @@
+export const INCREMENT = 'INCREMENT'
+export const DECREMENT = 'DECREMENT'
